+++ library/src/main/java/com/google/android/exoplayer/upstream/DataSourceInputStream.java
@@ -43,6 +43,19 @@ public class DataSourceInputStream extends InputStream {
     singleByteArray = new byte[1];
   }
 
+  /**
+   * Optional call to open the underlying {@link DataSource}.
+   * <p>
+   * Calling this method does nothing if the {@link DataSource} is already open. Calling this
+   * method is optional, since the read and skip methods will automatically open the underlying
+   * {@link DataSource} if it's not open already.
+   *
+   * @throws IOException If an error occurs opening the {@link DataSource}.
+   */
+  public void open() throws IOException {
+    checkOpened();
+  }
+
   @Override
   public int read() throws IOException {
     read(singleByteArray);
