+++ library/src/main/java/com/google/android/exoplayer/upstream/NetworkLoadable.java
@@ -13,15 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.exoplayer.util;
+package com.google.android.exoplayer.upstream;
 
 import com.google.android.exoplayer.ParserException;
 import com.google.android.exoplayer.upstream.Loader.Loadable;
 
+import android.net.Uri;
+
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.URL;
-import java.net.URLConnection;
 
 /**
  * A {@link Loadable} for loading an object over the network.
@@ -40,21 +40,16 @@ public final class NetworkLoadable<T> implements Loadable {
      *
      * @param connectionUrl The source of the response, after any redirection.
      * @param inputStream An {@link InputStream} from which the response data can be read.
-     * @param inputEncoding The encoding of the data, if available.
      * @return The parsed object.
      * @throws ParserException If an error occurs parsing the data.
      * @throws IOException If an error occurs reading data from the stream.
      */
-    T parse(String connectionUrl, InputStream inputStream, String inputEncoding)
-        throws ParserException, IOException;
+    T parse(String connectionUrl, InputStream inputStream) throws ParserException, IOException;
 
   }
 
-  public static final int DEFAULT_TIMEOUT_MILLIS = 10000;
-
-  private final String url;
-  private final String userAgent;
-  private final int timeoutMillis;
+  private final DataSpec dataSpec;
+  private final HttpDataSource httpDataSource;
   private final Parser<T> parser;
 
   private volatile T result;
@@ -62,24 +57,13 @@ public final class NetworkLoadable<T> implements Loadable {
 
   /**
    * @param url The url from which the object should be loaded.
-   * @param userAgent The user agent to use when requesting the object.
+   * @param httpDataSource A {@link HttpDataSource} to use when loading the data.
    * @param parser Parses the object from the network response.
    */
-  public NetworkLoadable(String url, String userAgent, Parser<T> parser) {
-    this(url, userAgent, DEFAULT_TIMEOUT_MILLIS, parser);
-  }
-
-  /**
-   * @param url The url from which the object should be loaded.
-   * @param userAgent The user agent to use when requesting the object.
-   * @param timeoutMillis The desired http timeout in milliseconds.
-   * @param parser Parses the object from the network response.
-   */
-  public NetworkLoadable(String url, String userAgent, int timeoutMillis, Parser<T> parser) {
-    this.url = url;
-    this.userAgent = userAgent;
-    this.timeoutMillis = timeoutMillis;
+  public NetworkLoadable(String url, HttpDataSource httpDataSource, Parser<T> parser) {
+    this.httpDataSource = httpDataSource;
     this.parser = parser;
+    dataSpec = new DataSpec(Uri.parse(url));
   }
 
   /**
@@ -103,28 +87,13 @@ public final class NetworkLoadable<T> implements Loadable {
 
   @Override
   public final void load() throws IOException, InterruptedException {
-    String inputEncoding;
-    InputStream inputStream = null;
+    DataSourceInputStream inputStream = new DataSourceInputStream(httpDataSource, dataSpec);
     try {
-      URLConnection connection = configureConnection(new URL(url));
-      inputStream = connection.getInputStream();
-      inputEncoding = connection.getContentEncoding();
-      result = parser.parse(connection.getURL().toString(), inputStream, inputEncoding);
+      inputStream.open();
+      result = parser.parse(httpDataSource.getUrl(), inputStream);
     } finally {
-      if (inputStream != null) {
-        inputStream.close();
-      }
+      inputStream.close();
     }
-  }
-
-  private URLConnection configureConnection(URL url) throws IOException {
-    URLConnection connection = url.openConnection();
-    connection.setConnectTimeout(timeoutMillis);
-    connection.setReadTimeout(timeoutMillis);
-    connection.setDoOutput(false);
-    connection.setRequestProperty("User-Agent", userAgent);
-    connection.connect();
-    return connection;
   }
 
 }
